feat(plugin): adapter plugin -- dameng
