+++ plugin/datacap-plugin-dm/src/main/java/io/edurt/datacap/plugin/DmPlugin.java
@@ -1,0 +1,5 @@
+package io.edurt.datacap.plugin;
+
+public class DmPlugin
+        extends Plugin
+{}
