+++ pom.xml
@@ -36,7 +36,7 @@
         <module>plugin/datacap-plugin-h2</module>
         <module>plugin/datacap-plugin-hdfs</module>
         <module>plugin/datacap-plugin-db2</module>
-        <!--        <module>plugin/datacap-jdbc-dm</module>-->
+        <module>plugin/datacap-plugin-dm</module>
         <!--        <module>plugin/datacap-jdbc-dremio</module>-->
         <!--        <module>plugin/datacap-jdbc-druid</module>-->
         <!--        <module>plugin/datacap-jdbc-duckdb</module>-->
